TRQ-3798. Make our decimal division for cgroup memory more precise.
Conflicts:
	src/test/numa_chip/scaffolding.c

+++ src/test/numa_socket/scaffolding.c
@@ -258,7 +258,7 @@ int get_machine_total_memory(hwloc_topology_t topology, unsigned long *memory)
   return(PBSE_NONE);
   }
 
-float Chip::how_many_tasks_fit(req const &r, int place_type) const
+double Chip::how_many_tasks_fit(req const &r, int place_type) const
   {
   return(tasks);
   }
